+++ blagominer/inout.nogui.h
@@ -29,7 +29,7 @@ public:
 	) override;
 
 	void printPlotsStart() override;
-	void printPlotsInfo(char const* const directory, unsigned nfiles, unsigned long long size) override;
+	void printPlotsInfo(char const* const directory, size_t nfiles, unsigned long long size) override;
 	void printPlotsEnd(unsigned long long total_size) override;
 
 	void printThreadActivity(
@@ -50,7 +50,7 @@ public:
 		unsigned long long deadline
 	) override;
 	void printScanProgress(
-		int ncoins, std::wstring const& connQualInfo,
+		size_t ncoins, std::wstring const& connQualInfo,
 		unsigned long long bytesRead, unsigned long long round_size,
 		double thread_time, double threads_speed,
 		unsigned long long deadline
@@ -99,7 +99,7 @@ public:
 		unsigned long long targetDeadline
 	) override;
 	void printConnQuality(
-		int ncoins, std::wstring const& connQualInfo
+		size_t ncoins, std::wstring const& connQualInfo
 	) override;
 
 	void debugRoundTime(
